requirements.txt
praw~=7.7.1
spacy~=4.0.0.dev2
mysql-connector-python~=8.3.0
